added spring-boot-security to enable logging

+++ src/main/java/pl/sk8/trickslist/current_trick/CurrentTrickRepository.java
@@ -2,6 +2,6 @@ package pl.sk8.trickslist.current_trick;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CurrentTrickRepository extends JpaRepository<CurrentTrick, Long> {
+public interface CurrentTrickRepository extends JpaRepository<CurrentTrick, Integer> {
 
 }
