+++ src/main/java/pl/sk8/trickslist/difficulty/DifficultyRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface DifficultyRepository extends JpaRepository<Difficulty, Long> {
+public interface DifficultyRepository extends JpaRepository<Difficulty, Integer> {
     Optional<Difficulty> findById(int id);
     Optional<Difficulty> findByName(String name);
 }
